fix sanity checking
Simpler version of:

  commit 107b4c81a10d18a1faff77545765080e64d98f22
  Author: YomikoR <monommino@gmail.com>
  Date:   Wed Sep 22 19:04:20 2021 +0800

      file drop: improved sanity check

+++ vsedit/src/script_templates/drop_file_category_model.cpp
@@ -240,9 +240,9 @@ void DropFileCategoryModel::addCategory()
 // END OF void DropFileCategoryModel::addCategory()
 //==============================================================================
 
-void DropFileCategoryModel::deleteCategory(int a_index)
+void DropFileCategoryModel::deleteCategory(unsigned a_index)
 {
-    if (a_index >= (int)m_categories.size()) {
+    if (a_index >= m_categories.size()) {
         return;
     }
 
@@ -254,9 +254,9 @@ void DropFileCategoryModel::deleteCategory(int a_index)
 // END OF void DropFileCategoryModel::deleteCategory(int a_index)
 //==============================================================================
 
-QString DropFileCategoryModel::sourceTemplate(int a_index) const
+QString DropFileCategoryModel::sourceTemplate(unsigned a_index) const
 {
-    if (a_index >= (int)m_categories.size()) {
+    if (a_index >= m_categories.size()) {
         return QString();
     }
 
@@ -266,10 +266,10 @@ QString DropFileCategoryModel::sourceTemplate(int a_index) const
 // END OF QString DropFileCategoryModel::sourceTemplate(int a_index) const
 //==============================================================================
 
-void DropFileCategoryModel::setSourceTemplate(int a_index,
+void DropFileCategoryModel::setSourceTemplate(unsigned a_index,
         const QString &a_text)
 {
-    if (a_index >= (int)m_categories.size()) {
+    if (a_index >= m_categories.size()) {
         return;
     }
